% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{density_normal}
\alias{density_normal}
\title{Multivariate normal probability density function}
\usage{
density_normal(x, mu, Sigma_inv, logret = FALSE)
}
\arguments{
\item{x}{A numeric matrix, each column of which is a point at which the
density is to be evaluated.}

\item{mu}{A numeric vector, the mean of the distribution.}

\item{Sigma_inv}{A numeric matrix, the precision matrix (inverse of the
of the variance-covariance matrix) of the distribution.}

\item{logret,}{a logical value indicating whether to return the log density.}
}
\value{
A column vector whose jth element is the density of a multivariate
normal distribution with mean mu and precision matrix Sigma_inv evaluated at
the jth column of x. If logret is true, the natural logarithm of the density
is returned.
}
\description{
Multivariate normal probability density function
}
\examples{
M <- matrix(c(1, 0.5, 0.5, 1), 2, 2)
m <- c(0, 0)
density_normal(cbind(c(0, 0), c(2, 2)), m, solve(M))
density_normal(cbind(c(0, 0), c(2, 2)), m, solve(M), TRUE)
}

