% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{density_t}
\alias{density_t}
\title{Multivariate Student-t probability density function}
\usage{
density_t(x, nu, mu, Sigma_inv, logret = FALSE)
}
\arguments{
\item{x}{A numeric matrix, each column of which is a point at which the
density is to be evaluated.}

\item{nu}{A positive integer, the degrees of freedom of the distribution.}

\item{mu}{A numeric vector, the location parameter of the distribution.}

\item{Sigma_inv}{A numeric matrix, the inverse of the scale matrix the
of the variance-covariance matrix) of the distribution.}

\item{logret, }{a logical value indicating whether to return the log density.
Defaults to FALSE.}
}
\value{
A column vector whose jth element is the density of a multivariate
Student-t with the specified parameters evaluated at  the jth column of x.
If logret is true, the natural logarithm of the density is returned.
}
\description{
Multivariate Student-t probability density function
}
\examples{
M <- matrix(c(1, 0.5, 0.5, 1), 2, 2)
m <- c(0, 0)
df <- 20
density_t(cbind(c(0, 0), c(2, 2)), df, m, solve(M))
density_t(cbind(c(0, 0), c(2, 2)), df, m, solve(M), TRUE)
}

