% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{density_wishart}
\alias{density_wishart}
\title{Wishart probability density function}
\usage{
density_wishart(X, v, S, logret = FALSE)
}
\arguments{
\item{v}{An integer, the degrees of freedom of the distribution.}

\item{S}{A numeric matrix, the scale matrix of the distribution.
of the variance-covariance matrix) of the distribution.}

\item{logret,}{a logical value indicating whether to return the log density.}

\item{x}{A numeric matrix, the point at which the density is to be
evaluated.}
}
\value{
A column real number: the value of the probability density function
by the default or the natural logarithm if logret is TRUE.
}
\description{
Wishart probability density function
}
\examples{
M <- matrix(c(1, 0.5, 0.5, 1), 2, 2)
density_wishart(M, 10, M)
density_wishart(M, 10, M, TRUE)
}

