% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_gamma}
\alias{draw_gamma}
\title{Simulate draws from the Gamma distribution.}
\usage{
draw_gamma(a, rate)
}
\arguments{
\item{a}{A positive number, the shape parameter of the distribution.}

\item{rate}{A numeric vector of positive values containing the rates
rates parameters for each draw.}
}
\value{
A vector of random draws. The jth element is a Gamma(a, rate(j))
variate.
}
\description{
Simulate draws from the Gamma distribution.
}
\details{
Note that the rate parameter is the reciprocal of the scale
parameter which appears as an argument in the underlying C function rgamma
from Rmath.h called here. The function draw_gamma is vectorized for its
second argument, the vector of rate parameters, and uses the length of this
parameter to determine how many draws to make. This function is intended
for use in a Gibbs sampler that uses the scale mixture of normals
representation of the Student-t distribution. In each step of the sampler
we draw an auxiliary parameter (a Gamma variate) for each observation in the
sample. Each of these draws has the same shape parameter but a different
rate parameter.
}
\examples{
draw_gamma(1, 1:10)
}

