% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_normal}
\alias{draw_normal}
\title{Simulate one draw from a multivariate normal distribution}
\usage{
draw_normal(mu, Sigma_inv)
}
\arguments{
\item{mu}{A numeric vector, the mean of the distribution.}

\item{Sigma_inv}{A numeric matrix, the precision matrix (inverse of the
of the variance-covariance matrix) of the distribution}
}
\value{
A column vector containing the draw.
}
\description{
Simulate one draw from a multivariate normal distribution
}
\details{
Uses the Cholesky decomposition of Sigma_inv.
}
\examples{
M <- matrix(c(1, 0.5, 0.5, 1), 2, 2)
draw_normal(c(0, 0), solve(M))
}

