% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_wishart}
\alias{draw_wishart}
\title{Simulate one draw from the Wishart distribution}
\usage{
draw_wishart(v, S)
}
\arguments{
\item{v}{An integer, the degrees of freedom of the distribution.}

\item{S}{A numeric matrix, the scale matrix of the distribution.}
}
\value{
A column vector containing the draw.
}
\description{
Simulate one draw from the Wishart distribution
}
\details{
Employs the Bartlett Decomposition (Smith & Hocking 1972).
Output exactly matches that of rwish from the MCMCpack package if the same
random seed is used.
}
\examples{
M <- matrix(c(1, 0.5, 0.5, 1), 2, 2)
draw_wishart(10, M)
}

