% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ff5}
\alias{ff5}
\title{Fama/French 5 Factors (2x3)}
\format{\preformatted{'data.frame':	622 obs. of  7 variables:
 $ month : Factor w/ 622 levels "196307","196308",..: 1 2 3 4 5 6 7 8 9 10 ...
 $ Mkt.RF: num  -0.39 5.07 -1.57 2.53 -0.85 1.83 2.24 1.54 1.41 0.1 ...
 $ SMB   : num  -0.46 -0.85 -0.5 -1.3 -0.83 -1.87 0.12 0.3 1.43 -1.52 ...
 $ HML   : num  -0.82 1.63 0.19 -0.11 1.66 -0.11 1.64 2.83 3.36 -0.42 ...
 $ RMW   : num  0.72 0.42 -0.8 2.75 -0.34 0.18 0.21 0.04 -2.01 -1.39 ...
 $ CMA   : num  -1.15 -0.33 0.27 -2.24 2.22 -0.3 1.46 0.86 2.93 -1.08 ...
 $ RF    : num  0.27 0.25 0.27 0.29 0.27 0.29 0.3 0.26 0.31 0.29 ...
}}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/
data_library.html}
}
\usage{
ff5
}
\description{
Monthly observations for the Fama & French (2015) five-factor
model.
}
\references{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/
Data_Library/f-f_5_factors_2x3.html}
}
\keyword{datasets}

