% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{industry10}
\alias{industry10}
\title{10 Industry Portfolios}
\format{An object of class \code{list} of length 2.}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/
data_library.html}
}
\usage{
industry10
}
\description{
Monthly returns for ten portfolios formed by industry.
}
\references{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/
Data_Library/det_10_ind_port.html}
}
\keyword{datasets}

