% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{liq}
\alias{liq}
\title{Liquidity Factor}
\format{An object of class \code{data.frame} with 629 rows and 4 columns.}
\source{
\url{http://faculty.chicagobooth.edu/lubos.pastor/research/}
}
\usage{
liq
}
\description{
Monthly observations for the liquidity factor of Pastor &
Stambaugh (2003).
}
\references{
\url{http://faculty.chicagobooth.edu/lubos.pastor/research/
liq_data_1962_2014.txt}
}
\keyword{datasets}

