% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log1p_arma}
\alias{log1p_arma}
\title{Armadillo wrapper for R's log1p function.}
\usage{
log1p_arma(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector whose values are log(1 + x).
}
\description{
Armadillo wrapper for R's log1p function.
}
\details{
This is simply a wrapper to the vectorized Rcpp sugar function
log1p which uses the same implementation as R. It takes an Armadillo vector
as input and returns and Armadillo vector as output. It is intended for use
in C++ code that uses only Armadillo types.
}
\examples{
log1p_arma(1/10^(0:5))
}

