% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_mv_gamma}
\alias{log_mv_gamma}
\title{Natural logarithm of the p-dimensional MV Gamma function}
\usage{
log_mv_gamma(p, a)
}
\arguments{
\item{p}{An integer, the dimesion of the MV Gamma function.}

\item{a}{A real number, the argument of the MV Gamma function.}
}
\description{
Natural logarithm of the p-dimensional MV Gamma function
}
\details{
Used to calculate the normalizing constant for the density of the
Wishart distribution.
}

