% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{mom}
\alias{mom}
\title{Momentum Factor}
\format{An object of class \code{data.frame} with 1064 rows and 2 columns.}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/
data_library.html}
}
\usage{
mom
}
\description{
Monthly observations for the momentum factor.
}
\references{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/
Data_Library/det_mom_factor.html}
}
\keyword{datasets}

