% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{size}
\alias{size}
\title{Portfolios Formed on Size}
\format{An object of class \code{list} of length 2.}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/
data_library.html}
}
\usage{
size
}
\description{
Monthly returns for portfolios formed by quantiles of size.
}
\references{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/
Data_Library/det_port_form_sz.html}
}
\keyword{datasets}

