% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ff5}
\alias{ff5}
\title{Fama/French 5 Factors (2x3)}
\format{An object of class \code{data.frame} with 626 rows and 7 columns.}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/
data_library.html}
}
\usage{
ff5
}
\description{
Monthly observations for the Fama & French (2015) five-factor
model.
}
\references{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/
Data_Library/f-f_5_factors_2x3.html}
}
\keyword{datasets}

