% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{industry10}
\alias{industry10}
\title{10 Industry Portfolios}
\format{\preformatted{List of 2
 $ equal:'data.frame':	1062 obs. of  11 variables:
  ..$ month: Factor w/ 1062 levels "192607","192608",..: 1 2 3 4 5 6 7 8 9 10 ...
  ..$ NoDur: num [1:1062] 1.17 5.28 0.03 -2.59 6.07 0.41 -1.49 3.58 -3.23 1.33 ...
  ..$ Durbl: num [1:1062] 4.03 0.28 -3.34 -9.46 -0.38 4.8 -2.27 6.06 -2.36 -1.17 ...
  ..$ Manuf: num [1:1062] 2.3 2.14 0.53 -5.11 1.69 3.39 0.78 6.38 -1.24 1.57 ...
  ..$ Enrgy: num [1:1062] -2.19 4.36 -3.51 2.67 1.59 4.92 3.68 4.24 -9.15 -3.93 ...
  ..$ HiTec: num [1:1062] 2.21 2.61 0.86 -5.72 1.32 1.46 0.44 8.33 -1.58 2.04 ...
  ..$ Telcm: num [1:1062] 1.31 1.08 0.22 -0.99 1.48 1.62 0.13 6.98 1.35 0.58 ...
  ..$ Shops: num [1:1062] 0.69 -0.03 -0.71 -3.28 1.35 2.81 -2.29 5.1 0.25 3.23 ...
  ..$ Hlth : num [1:1062] 2.45 6.11 0.8 0.13 7.2 0.42 6.13 2.47 -0.2 5.8 ...
  ..$ Utils: num [1:1062] 4.85 -2 2.06 -2.98 5.71 1.72 -0.44 4.88 2.19 1.69 ...
  ..$ Other: num [1:1062] 0.21 4.53 -0.53 -2.51 0.85 1.23 5.16 7.81 0.04 1.44 ...
 $ value:'data.frame':	1062 obs. of  11 variables:
  ..$ month: Factor w/ 1062 levels "192607","192608",..: 1 2 3 4 5 6 7 8 9 10 ...
  ..$ NoDur: num [1:1062] 1.45 3.97 1.14 -1.24 5.21 0.82 -0.67 3.42 2.74 3.36 ...
  ..$ Durbl: num [1:1062] 15.55 3.68 4.8 -8.23 -0.19 ...
  ..$ Manuf: num [1:1062] 4.69 2.81 1.15 -3.63 4.1 3.74 -0.08 5.81 1.43 0.77 ...
  ..$ Enrgy: num [1:1062] -1.18 3.37 -3.39 -0.78 0.02 2.82 1.29 1.48 -6.02 -5.17 ...
  ..$ HiTec: num [1:1062] 2.9 2.66 -0.38 -4.58 4.71 -0.02 -1.13 4.45 1.45 5.4 ...
  ..$ Telcm: num [1:1062] 0.83 2.17 2.41 -0.11 1.63 1.99 1.88 3.97 5.56 -2.13 ...
  ..$ Shops: num [1:1062] 0.11 -0.71 0.21 -2.29 6.43 0.62 -2.55 3.61 -0.41 4.43 ...
  ..$ Hlth : num [1:1062] 1.77 4.25 0.69 -0.57 5.42 0.11 5.05 1.71 1.01 2.74 ...
  ..$ Utils: num [1:1062] 7.04 -1.69 2.04 -2.63 3.71 -0.17 -1.79 4.53 0.37 1.71 ...
  ..$ Other: num [1:1062] 2.16 4.38 0.29 -2.85 2.11 3.4 1.51 5.06 1.27 0.86 ...
}}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/
data_library.html}
}
\usage{
industry10
}
\description{
Monthly returns for ten portfolios formed by industry.
}
\references{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/
Data_Library/det_10_ind_port.html}
}
\keyword{datasets}

