% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{liq}
\alias{liq}
\title{Liquidity Factor}
\format{\preformatted{'data.frame':	629 obs. of  4 variables:
 $ month   : int  196208 196209 196210 196211 196212 196301 196302 196303 196304 196305 ...
 $ LIQagg  : num  -0.01758 -0.00409 -0.10423 -0.01962 -0.00529 ...
 $ LIQinnov: num  0.00437 0.01305 -0.07302 0.03037 0.01429 ...
 $ LIQv    : num  NA NA NA NA NA NA NA NA NA NA ...
}}
\source{
\url{http://faculty.chicagobooth.edu/lubos.pastor/research/}
}
\usage{
liq
}
\description{
Monthly observations for the liquidity factor of Pastor &
Stambaugh (2003).
}
\references{
\url{http://faculty.chicagobooth.edu/lubos.pastor/research/
liq_data_1962_2014.txt}
}
\keyword{datasets}

