% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{mom}
\alias{mom}
\title{Momentum Factor}
\format{\preformatted{'data.frame':	1060 obs. of  2 variables:
 $ month: Factor w/ 1060 levels "192701","192702",..: 1 2 3 4 5 6 7 8 9 10 ...
 $ MOM  : num  0.44 -2.01 3.59 4.19 3.01 0.51 4.32 1.12 1.93 -1.11 ...
}}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/
data_library.html}
}
\usage{
mom
}
\description{
Monthly observations for the momentum factor.
}
\references{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/Data_Library/det_mom_factor.html}
}
\keyword{datasets}

