% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{size}
\alias{size}
\title{Portfolios Formed on Size}
\format{\preformatted{List of 2
 $ equal:'data.frame':	1066 obs. of  20 variables:
  ..$ month: int [1:1066] 192607 192608 192609 192610 192611 192612 192701 192702 192703 192704 ...
  ..$ <=0  : logi [1:1066] NA NA NA NA NA NA ...
  ..$ Lo30 : num [1:1066] -0.43 3.9 -1.08 -3.32 -0.46 2.54 1.86 6.45 -3.06 3.48 ...
  ..$ Med40: num [1:1066] 1.52 3.04 -0.54 -3.52 3.82 1.69 1.32 6.62 -1.93 -0.15 ...
  ..$ Hi30 : num [1:1066] 2.68 2.09 0.16 -3.06 3.09 3.25 0.23 4.17 -0.61 0.05 ...
  ..$ Lo20 : num [1:1066] -0.57 3.84 -0.48 -3.29 -0.55 1.19 2.61 7.03 -3.7 4.85 ...
  ..$ Qnt2 : num [1:1066] 0.59 3.59 -1.4 -4.1 2.18 3.78 1.26 5.68 -1.52 0.02 ...
  ..$ Qnt3 : num [1:1066] 1.6 3.71 0 -2.89 3.41 1.32 0.74 6.83 -3.26 0.11 ...
  ..$ Qnt4 : num [1:1066] 1.47 1.61 -0.5 -3.36 3.39 2.63 1.22 5.34 -1.09 -0.16 ...
  ..$ Hi20 : num [1:1066] 3.33 2.33 -0.09 -2.95 3.16 3.14 -0.08 4.28 0.19 0.14 ...
  ..$ Lo10 : num [1:1066] -1.45 5.12 0.93 -4.84 -0.78 -2.66 4.53 6.85 -5.91 4.67 ...
  ..$ Dec2 : num [1:1066] 0.29 2.59 -1.87 -1.77 -0.32 4.96 0.77 7.19 -1.65 5.01 ...
  ..$ Dec3 : num [1:1066] -0.15 4.03 -2.27 -3.36 -0.29 5.25 0.34 5.31 -1.79 0.68 ...
  ..$ Dec4 : num [1:1066] 1.33 3.15 -0.53 -4.83 4.65 2.31 2.18 6.05 -1.24 -0.62 ...
  ..$ Dec5 : num [1:1066] 1.24 2.72 0.07 -2.98 3.24 1.35 -0.64 5.63 -3.14 0.43 ...
  ..$ Dec6 : num [1:1066] 1.98 4.72 -0.07 -2.8 3.57 1.28 2.15 8.06 -3.39 -0.23 ...
  ..$ Dec7 : num [1:1066] 1.55 1.6 -1.64 -3.45 3.82 1.82 1.61 6.75 0.02 -0.2 ...
  ..$ Dec8 : num [1:1066] 1.38 1.63 0.64 -3.27 2.95 3.45 0.83 3.94 -2.19 -0.12 ...
  ..$ Dec9 : num [1:1066] 3.38 0.98 -0.86 -3.47 3.61 3.27 0.27 4.91 -0.46 -0.61 ...
  ..$ Hi10 : num [1:1066] 3.29 3.7 0.67 -2.43 2.7 3.01 -0.43 3.65 0.85 0.9 ...
 $ value:'data.frame':	1066 obs. of  20 variables:
  ..$ month: int [1:1066] 192607 192608 192609 192610 192611 192612 192701 192702 192703 192704 ...
  ..$ <=0  : logi [1:1066] NA NA NA NA NA NA ...
  ..$ Lo30 : num [1:1066] 0.14 3.21 -1.74 -2.94 -0.38 4.15 1.21 6.41 -1.96 3.8 ...
  ..$ Med40: num [1:1066] 1.59 2.77 -0.88 -3.28 3.73 1.51 1.59 6.95 -1.51 0.3 ...
  ..$ Hi30 : num [1:1066] 3.42 2.91 0.8 -2.79 2.74 3.04 -0.03 4.2 0.92 0.84 ...
  ..$ Lo20 : num [1:1066] 0.37 2.26 -1.39 -2.57 -0.95 2.46 1.53 7.65 -3.57 5.07 ...
  ..$ Qnt2 : num [1:1066] 0.78 3.51 -1.06 -3.93 2.94 3.58 2.14 5.86 -0.77 1.25 ...
  ..$ Qnt3 : num [1:1066] 1.68 3.75 0.05 -2.67 3.52 1.33 1.14 6.98 -3.12 0.47 ...
  ..$ Qnt4 : num [1:1066] 1.39 1.53 -0.26 -3.38 3.25 2.68 1.1 5.11 -1.16 0.04 ...
  ..$ Hi20 : num [1:1066] 3.67 3.07 0.81 -2.74 2.71 3 -0.12 4.22 1.24 0.93 ...
  ..$ Lo10 : num [1:1066] -0.12 1.33 0.59 -4.33 -3.3 -2.49 1.67 6.44 -5.46 5.23 ...
  ..$ Dec2 : num [1:1066] 0.52 2.55 -2 -2.01 -0.23 3.93 1.49 7.98 -3.05 5.03 ...
  ..$ Dec3 : num [1:1066] -0.05 4 -2.01 -3.25 0.08 5.6 0.95 5.35 -0.57 2.71 ...
  ..$ Dec4 : num [1:1066] 1.31 3.2 -0.46 -4.35 4.74 2.4 2.85 6.16 -0.88 0.42 ...
  ..$ Dec5 : num [1:1066] 1.21 2.81 -0.06 -2.93 3.64 1.21 -0.53 5.56 -3.04 0.69 ...
  ..$ Dec6 : num [1:1066] 2.04 4.45 0.14 -2.48 3.44 1.42 2.35 7.99 -3.17 0.31 ...
  ..$ Dec7 : num [1:1066] 1.58 1.61 -2.02 -3.6 3.63 1.41 1.67 7.18 0.02 0.08 ...
  ..$ Dec8 : num [1:1066] 1.29 1.49 0.74 -3.26 3.05 3.37 0.8 3.98 -1.82 0.02 ...
  ..$ Dec9 : num [1:1066] 3.53 0.61 -0.77 -3.36 3.86 3.11 0.43 4.9 -0.69 -0.18 ...
  ..$ Hi10 : num [1:1066] 3.71 3.79 1.25 -2.56 2.4 2.97 -0.27 4.03 1.78 1.24 ...
}}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/
data_library.html}
}
\usage{
size
}
\description{
Monthly returns for portfolios formed by quantiles of size.
}
\references{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/
Data_Library/det_port_form_sz.html}
}
\keyword{datasets}

